First pass at message refactoring
Add MessageFlags, which controls the formatter used. In the
future this will be the basis for differentiating between
self-sent lines (#101), highlights (#103), colouring different
types of event (#105) and per-event notifications (#192).

The current implementation is a bit ugly. I think it needs
pulling out of WindowModel and being cleaned up a bit.